add shared setup for jsdom
