+++ test-utils/index.ts
@@ -1,3 +1,6 @@
+import * as setupJsDomAll from './src/jsdom/setupJsDom';
+
+export const setupJsDom = setupJsDomAll.setupJsDom as (global?: any) => void;
 /*
  * Re-exports everything from @testing-library/react and extends it with:
  * 1) additional queries
