+++ test-utils/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "extends": "../tsconfig.json",
   "compilerOptions": {
+    "allowJs": true,
     "rootDir": ".",
     "outDir": "./build"
   },
